move computeReadiumCssJsonMessage in common
